tabs panel refactored
