+++ src/ui/tabs-panel/tabs-panel.tsx
@@ -5,19 +5,26 @@ import { useTabsPanel } from './tabs-panel.hook'
 import { TabDropdown } from '../tab-dropdown'
 import { SortableContainer, SortableElement } from 'react-sortable-hoc'
 import { TabData } from './tabs-panel.types'
+import { useSortable } from './tabs-panel.hook.sortable'
+import { useContextMenu } from './tabs-panel.hook.contextmenu'
+import { useDropdown } from './tabs-panel.hook.dropdown'
+import { useRequestHandlers } from './tabs-panel.hook.request'
+import { useTabs } from './tabs-panel.hook.tabs'
 
 const id = Math.random().toString(36).substring(2)
 export const TabsPanel: FC = () => {
   const {
     tabs,
-    collapsedTabs,
-    onTabClicked,
-    onCloseClicked,
-    onCloseOthersClicked,
-    onClickCloseOnRight,
-    onClickCloseAll,
-    onSortEnd,
-  } = useTabsPanel(id)
+    setTabs,
+    tabDataRef,
+  } = useTabsPanel()
+
+  const { onSortEnd } = useSortable({ tabs, setTabs })
+  const { collapsedTabs } = useDropdown({ tabs, id })
+  const { onCloseClicked, onTabClicked, showTabContent } = useTabs({ tabs, setTabs })
+  const { onClickCloseAll, onClickCloseOnRight, onCloseOthersClicked } = useContextMenu({ tabs, setTabs, showTabContent })
+
+  useRequestHandlers({ setTabs, tabDataRef })
 
   const SortableItem = SortableElement(({ tab }: { tab: TabData }) => <TabButton
     onClickButton={() => onTabClicked(tab.requestId)}
